cpu_input_boost: add module/runtime tunables
 - replace all kconfig/build-time options with module/runtime tunables
 - make fb_boost && input_boost have different paramaters
 - change cpu-boost's input_boost_ms default value.

NOTE: I'm adding params for a number of reasons. examples;

 - runtime re-configuration for testing purposes
 - runtime configuration allows creating 'switchable' profiles
 - the parameter values should be easily checked/viewed

Also, I haven't yet wired up all of the types of boosting events that I plan
on supporting. e.g: much like the fb_boost ~ I'd like to have a dsb_audio_boost,
specifically meant for the low latency and/or aaudio use-cases.

due to the nature of the types of events that will be tunable; profiles could be
leveraged beyond just power-saving / energy efficiency or performace. ~ it will
be possible to tune this driver based on the type of workload.

Signed-off-by: ninez <johnstonljordan@gmail.com>